Muuda float double'itekš

+++ AndmeSalvestaja/CAS.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using System.IO;
 using Newtonsoft.Json;
 
-using VecT = System.Collections.Generic.List<System.Tuple<System.DateTime, float>>;
+using VecT = System.Collections.Generic.List<System.Tuple<System.DateTime, double>>;
 
 namespace AndmeSalvestaja
 {
